// client/src/types.ts
export interface BaseConversationNode {
  node_type: ibis_node_type_type;
  status: publication_state_type;
  created_at: string;
  published_at: string;
  updated_at: string;
  title: string;
  description: string;
  url: string;
  meta: meta_state_type;
  draft_for_role_id: number;
}

export interface ConversationNode extends BaseConversationNode {
  id: number;
  quest_id: number;
  guild_id?: number;
  creator_id: number;
  parent_id?: number;
  ancestry: string;
}

export const conversationNodePatchKeys: KeyArray<ConversationNode> = [
  "id",
  "quest_id",
  "guild_id",
  "creator_id",
  "parent_id",
  "node_type",
  "status",
  "title",
  "description",
  "url",
  "meta",
  "draft_for_role_id",
];

export interface QTreeNode extends ConversationNode {
  // id: number | string;
  children?: QTreeNode[];
  label: string;
  color?: string;
  icon?: string;
  parent?: QTreeNode;
}

export interface ReadStatus {
  node_id: number;
  member_id: number;
  seconds_shown: number;
  status: boolean;
}

export interface ReadStatusData extends ReadStatus {
  node_count: number;
  read_count: number;
}

export type generic_id = number | string;
export interface PseudoNode extends BaseConversationNode {
  id: generic_id;
  quest_id: generic_id;
  guild_id?: generic_id;
  creator_id: generic_id;
  parent_id?: generic_id;
  children?: PseudoNode[];
}

export type MaybeRealNode = PseudoNode | QTreeNode;